% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{cv.lsm}
\alias{cv.lsm}
\title{Cross-Validation MSE}
\usage{
cv.lsm(
  z,
  d,
  anchorpoints,
  epsilon,
  n_angles,
  n_intervals,
  kernel_id,
  id,
  initial.position,
  lower.bound = rep(1e-08, length(initial.position)),
  upper.bound = c(c(Inf, Inf, pi/2), rep(Inf, length(initial.position) - 3)),
  lower.delta = 0.1,
  upper.delta = 10,
  n_threads = -1
)
}
\arguments{
\item{z}{the vector containing z(d)}

\item{d}{the matrix containing the coordinates in which we know the value of z}

\item{anchorpoints}{a matrix with the coordinates of the anchor points which can be obtained calling find_anchorpoints.lsm}

\item{epsilon}{the value of the bandwidth parameter epsilon}

\item{n_angles}{the number of angles for the grid}

\item{n_intervals}{the number of intervals for the grid}

\item{kernel_id}{the type of kernel to be used}

\item{id}{the type of variogram to be used}

\item{initial.position}{the starting position to be given to the optimizer}

\item{lower.bound}{the lower bound for the optimization, by default (1e-8, 1e-8, ...)}

\item{upper.bound}{the upper bound for the optimization, by default (Inf, Inf, pi/2, Inf, Inf, ...)}

\item{lower.delta}{set the minimum value for Cross-Validation search for optimal delta in smoothing equal to lowerdelta*epsilon}

\item{upper.delta}{set the maximum value for Cross-Validation search for optimal delta in smoothing equal to upperdelta*epsilon}

\item{n_threads}{the number of threads for OpenMP, by default is equal to -1, which means that OpenMP will use all the available threads.}
}
\description{
calculate the mean squared error via cross-validation
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
cv.lsm(y,d,a$anchorpoints,350,8,8,"gaussian","exponential", c(200,200,0.01,100))
}
