% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperFunctions.R
\name{find_anchorpoints.lsm}
\alias{find_anchorpoints.lsm}
\title{Find Anchor Points}
\usage{
find_anchorpoints.lsm(dataset, n, plot_output = TRUE)
}
\arguments{
\item{dataset}{a dataset full of coordinates}

\item{n}{a parameter proportional to the number of anchorpoints}

\item{plot_output}{if set to true plot the original dataset and the anchorpoints}
}
\value{
the coordinates of the anchor points found and the specification of the grid used to build them
}
\description{
given a dataset find the corresponding equally spaced anchorpoints
}
\details{
given a set of points this function builds n*n grid covering all the locations. Then takes as anchor points all the
centers of the cells such that at least one of the points of d belongs to the same cell
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12)
}
