% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotgrid}
\alias{plotgrid}
\title{Plot Grid}
\usage{
plotgrid(variogram, index)
}
\arguments{
\item{variogram}{an object of type "sample_variogram" built via variogram.lsm}

\item{index}{the index of the vector neighborhood for which segments are going to be plotted.}
}
\description{
A function to plot segments for all pairs of locations belonging to the same vector neighborhood.
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
vario <- variogram.lsm(y,d,a$anchorpoints,370,8,8,"gaussian")
plotgrid(vario, 0)
}
