% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotvario}
\alias{plotvario}
\title{Plot Vario}
\usage{
plotvario(variogram, pos)
}
\arguments{
\item{variogram}{a variogram object returned by variogram.lsm}

\item{pos}{the index of an anchor point where the empiric variogram is plotted}
}
\description{
plot all directional empiric variograms sequentially in n_angles number of directions, from -pi to pi in radians.
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
vario <- variogram.lsm(y,d,a$anchorpoints,370,8,8,"gaussian")
pltovario(vario, 0)
}
