% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperFunctions.R
\name{predict.lsm}
\alias{predict.lsm}
\title{Predict LSM (Kriging)}
\usage{
\method{predict}{lsm}(sol, newpos, plot_output = TRUE, print_output = TRUE, n_threads = -1)
}
\arguments{
\item{sol}{an object of type lsm obtained by calling findsolutions.lsm}

\item{newpos}{a matrix with the coordinates of the points where to evaluate z}

\item{plot_output}{if set to TRUE plot the solutions, by default is TRUE}

\item{print_output}{if set to FALSE suppress the console output, by default is TRUE}

\item{n_threads}{the number of threads for OpenMP, by default is equal to -1, which means that OpenMP will use all the available threads.}
}
\value{
an object containing the vector with the means, the vector with the punctual predictions and the vector with the kriging variance
in newpos
}
\description{
for each couple of coordinates in newpos predict the mean and punctual value of z
}
\details{
given an object of type "lsm" returned by findsolutions.lsm this function performs kriging on the coordinates provided by newpos
and possibly plot the results found
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
vario <- variogram.lsm(y,d,a$anchorpoints,370,8,8,"gaussian")
solu <- findsolutions.lsm(vario, "exponential", c(200,200,0.01,100))
previsions <- predict.lsm(solu, d)
}
