% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperFunctions.R
\name{smooth.lsm}
\alias{smooth.lsm}
\title{Smooth LSM}
\usage{
smooth.lsm(model, newpoints, n_threads = -1)
}
\arguments{
\item{model}{a "lsm" object generated via findsolutions.lsm}

\item{newpoints}{a matrix with the coordinates of the points the knowledge of the parameters is needed}

\item{n_threads}{the number of threads for OpenMP, by default is equal to -1, which means that OpenMP will use all the available threads.}
}
\value{
a matrix with the values of the paramters smoothed in newpoints
}
\description{
compute the value of the parameters of the variogram function of model in the points contained in newpoints
}
\details{
given model, this function exploits model$solutions and model$delta to perform smoothing and find the value of the 
parameters regulating the variogram function in other points beyond the anchor ones. model$delta already contains the optimal value of 
delta which does not need to be evaluated again.
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
vario <- variogram.lsm(y,d,a$anchorpoints,370,8,8,"gaussian")
solu <- findsolutions.lsm(vario, "exponential", c(200,200,0.01,100))
newparams <- smooth.lsm(solu, d)
}
