% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapperFunctions.R
\name{variogram.lsm}
\alias{variogram.lsm}
\title{Empiric Variogram LSM}
\usage{
variogram.lsm(
  z,
  d,
  anchorpoints,
  epsilon,
  n_angles,
  n_intervals,
  kernel_id,
  print_output = TRUE,
  n_threads = -1
)
}
\arguments{
\item{z}{the vector contatining f(d)}

\item{d}{the matrix contatining the coordinates in which we know the value of z}

\item{anchorpoints}{a matrix with the coordinates of the anchorpoints which can be obtained calling find_anchorpoints.lsm}

\item{epsilon}{the value of epsilon regulating the kernel}

\item{n_angles}{the number of angles for the grid}

\item{n_intervals}{the number of intervals for the grid}

\item{kernel_id}{the type of kernel to be used. At the moment the only possibility is "gaussian".}

\item{print_output}{if set to FALSE suppress the console output, by default is TRUE}

\item{n_threads}{the number of threads for OpenMP, by default is equal to -1, which means that OpenMP will use all the available threads.}
}
\value{
an object of type "sample_variogram" containing the kernel matrix, the grid matrix, the vecors with the value of x, y and norm of
every tile of the grid, the matrix of the squaredweights, the matrix with the sample variogam, the matrix with the anchor points used,
the value of the bandwidth parameter epsilon used, the id of the kernel function, the number of angles and of intervals used to build
the grid, the matrix with the coordinates of the initial points and the vector with the function z evaluated in these points.
}
\description{
compute the sample variogram in the anchorpoints
}
\details{
the purpose of this function is to calculate the value of the sample variogram in every anchor point. To do so 
the function requires to be given as input all the information about the construction of the grid and of the kernel as in the paper by
Fouedjio.
}
\examples{
data(meuse)
d <- cbind(meuse$x, meuse$y)
y <- meuse$elev
a <- find_anchorpoints.lsm(d,12,FALSE)
vario <- variogram.lsm(y,d,a$anchorpoints,370,8,8,"gaussian")
}
